Add files via upload
In my project, I implemented the Folium library to visualize trip directions on an interactive map. A distinctive red line on the map serves as a clear indicator of instances of overspeeding, providing users with an intuitive and actionable display of driving behavior. Folium's seamless integration with Python and Leaflet.js not only enhances the project's aesthetics but also improves user experience by allowing for easy navigation and detailed examination of specific locations. The result is a powerful tool that combines functionality with visual appeal, offering users insights into both trip directions and driving safety.